arch/ceva:use UP_DSB, UP_DMB, UP_ISB as barrier standard API
