+++ arch/ceva/include/spinlock.h
@@ -63,9 +63,6 @@
  *
  */
 
-#define UP_DSB() up_dsb()
-#define UP_DMB() up_dmb()
-
 /****************************************************************************
  * Public Types
  ****************************************************************************/
